+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Fixed a small visual glitch on Linux where the selected file was not
   highlighted in the file list in the thin or expanded file manager modes
 - Fixed heading ID links not working with, e.g., accents
+- Added "Copy filename", "Copy path" and "Copy ID" to document tabs context
+  menus
 
 ## Under the Hood